// src/packages/assistance/interfaces/userSofiaPlus.interface.ts
export interface IUserSofiaPlus {
  TIPO_DOCUMENTO?: string;
  DOCUMENTO: string;
  CORREO?: string;
  FICHA?: number;
  PROGRAMA_FORMACION?: string;
  CODIGO_CENTRO?: number;
  NOMBRE_CENTRO?: string;
  REGIONAL?: string;
  MODALIDAD_FORMACION?: string;
  TIPO_PROGRAMA?: string;
  ESTADO_APRENDIZ_EN_FICHA?: string;
}